refactor(dsg): move logging module
