+++ packages/data-service-generator/src/notify-build-manager.ts
@@ -1,5 +1,5 @@
 import { httpClient } from "./utils/http-client";
-import { defaultLogger as logger } from "./server/logging";
+import { defaultLogger as logger } from "./logging";
 
 interface BuildManagerNotifierOptions {
   buildManagerUrl: string;
