+++ packages/data-service-generator/src/server/resource/dto/create-enum-dto-module.ts
@@ -8,7 +8,7 @@ import {
   importDeclaration,
 } from "../../../utils/ast";
 import { createDTOFile } from "./create-dto-module";
-import { defaultLogger as logger } from "../../logging";
+import { defaultLogger as logger } from "../../../logging";
 
 const REGISTER_ENUM_TYPE_ID = builders.identifier("registerEnumType");
